+++ src/azure-cli/azure/cli/command_modules/role/custom.py
@@ -1106,14 +1106,11 @@ def delete_service_principal(cmd, identifier):
 
 def _get_app_object_id_from_sp_object_id(client, sp_object_id):
     sp = client.service_principals.get(sp_object_id)
-    app_object_id = None
+    result = list(client.applications.list(filter="appId eq '{}'".format(sp.app_id)))
 
-    if sp.service_principal_names:
-        result = list(client.applications.list(
-            filter="identifierUris/any(s:s eq '{}')".format(sp.service_principal_names[0])))
-        if result:
-            app_object_id = result[0].object_id
-    return app_object_id
+    if result:
+        return result[0].object_id
+    raise CLIError("Can't find associated application id from '{}'".format(sp_object_id))
 
 
 def list_service_principal_owners(cmd, identifier):